+++ tests/python_api/external_source_reader.py
@@ -42,8 +42,11 @@ def main():
 
     def image_dump(img, idx, device="cpu", mode=0):
         if device == "gpu":
-            import cupy as cp
-            img = cp.asnumpy(img)
+            try:
+                import cupy as cp
+                img = cp.asnumpy(img)
+            except ImportError:
+                pass
         img = img.transpose([1, 2, 0])  # NCHW
         img = (img).astype('uint8')
         if mode!=2: